% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAFlag.R
\name{AddAFlag}
\alias{AddAFlag}
\title{Add A Flag To A Subset Group}
\usage{
AddAFlag(SNPdataset, flagToAdd = "", taxa = c())
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{flagToAdd}{Unique flag for ingroup taxa}

\item{taxa}{Vector of names}
}
\value{
Returns a subset dataset with only ingroup taxa
}
\description{
This function will add a flag to the names taxa in the list.
}
\examples{
data(fakeData)
AddAFlag(fakeData, flagToAdd="WEST", taxa=c("in1", "in2", "in3"))
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ExportPops}
}

