% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addazero}
\alias{addazero}
\title{Add A Zero to alleleCounts list}
\usage{
addazero(alleleCounts, rownamesOriginal)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{rownamesOriginal}{rownames from the original full dataset.  These will be used to compare the subset to.}
}
\value{
Returns a new alleleCounts list with zeros in the right position
}
\description{
This function will add a zero if allele counts are missing from subpops
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ExportPops}
}

