% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNexDataToPhyData.R
\name{ConvertNexDataToPhyData}
\alias{ConvertNexDataToPhyData}
\title{Convert Nexus Data To Phylip Data}
\usage{
ConvertNexDataToPhyData(nexData)
}
\arguments{
\item{nexData}{Nexus formatted dataset}
}
\value{
Returns a phylip formated matrix
}
\description{
This function will convert data from nexus formatting into phylip. Used internally in the \code{ReadSNP} function.
}
\seealso{
\link{ReadSNP}
}

