% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOverlap.R
\name{DataOverlap}
\alias{DataOverlap}
\title{Calculate Amount of Data Overlap}
\usage{
DataOverlap(SNPdataset)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}
}
\value{
Returns a list of two items. The first item is a matrix of frequencies. The individual values are based on present/absent status. If a informational base (non-missing data) exists, then the frequency of present data is entered here.  For example, if all of the species in a site have data, then 1 will be entered for each one. Which base is present does not matter, only that there is or is not data present. This value is the same as the probability that the species contains data for that site.  The second item are the row/species means.  This value equates with the probability of an site overlapping with another species in the dataset.
}
\description{
This function will calculate the amount of data overlap in a dataset.  This assesses how much data is present in a dataset and how that data might be distributed across taxa.
}
\examples{
data(fakeData)
DataOverlap(fakeData)
}
\seealso{
\link{ReadSNP} \link{MakePresentAbsent}
}

