% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExportPops}
\alias{ExportPops}
\title{Export Population Allele Counts}
\usage{
ExportPops(SNPdataset, file = "", subsets = c(), includeFull = FALSE, ...)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{file}{A file name specified by either a variable or a double-quoted string. If not specified (ie, file=""), then  data is printed to the console.}

\item{subsets}{A vector of flags which will subset the populations.}

\item{includeFull}{logical, to include the full population data in addition to the subsets.}

\item{...}{additional arguments passed to \code{write.table}}
}
\value{
Returns a table with allele counts for a whole population and any subset populations.
}
\description{
This function will ...
}
\examples{
data(fakeData)
fakeData <- RemoveNonBinary(RemoveInvariantSites(fakeData))
fakeData <- AddAFlag(fakeData, flagToAdd="WEST", taxa=c("in1", "in2", "in3"))
ExportPops(fakeData, subsets=c("WEST"))
}
\seealso{
\link{ReadSNP} \link{AddAFlag} \link{ReturnAlleleCounts}
}

