% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetLinesToSkip}
\alias{GetLinesToSkip}
\title{Get Lines to Skip}
\usage{
GetLinesToSkip(file)
}
\arguments{
\item{file}{A file name specified by either a variable or a double-quoted string (ie, file="")}
}
\value{
Returns a number of lines to skip
}
\description{
This function will auto calculate the number of lines to skip before reading in either a phylip or nexus formatted file. It skips phylip-format summary lines and ones that are commented out or have ID tags (for example: ##, [ID: XX]).
}
\examples{
a <- matrix(data=c("AANNTGG", "AATTTGC", "TAAATGC"), dimnames=list(c("A", "B", "C"), "locus"))
WriteSNP(a, "example.phy")
GetLinesToSkip("example.phy")
}
\seealso{
\link{ReadSNP}
}

