% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberOfSitesForLocus.R
\name{GetNumberOfSitesForLocus}
\alias{GetNumberOfSitesForLocus}
\title{Get The Number Sites in a Locus}
\usage{
GetNumberOfSitesForLocus(SNPdataset, locus)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{locus}{Which locus to calculate}
}
\value{
Returns a number of sites
}
\description{
This function will calculate The number of sites for a given locus.
}
\examples{
data(fakeData)
GetNumberOfSitesForLocus(fakeData, 1)
GetNumberOfSitesForLocus(fakeData, 2)
GetNumberOfSitesForLocus(fakeData, 3)
}
\seealso{
\link{ReadSNP} \link{TakeSingleSNPfromEachLocus}
}
