% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IsMissing.R
\name{IsMissing}
\alias{IsMissing}
\title{Is Missing Data}
\usage{
IsMissing(locus, missingChar = "any")
}
\arguments{
\item{locus}{A single locus}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}
}
\value{
logical, return of whether is all missing data. For example, "AAAN" will be FALSE, "NNNN" will be TRUE.
}
\description{
This function will determine if data is all missing from a locus.
}
\examples{
data(fakeData)
IsMissing(fakeData[1,1])
IsMissing(fakeData[1,3])
}
\seealso{
\link{ReadSNP} \link{MakePresentAbsent} \link{CalculateMissingData}
}

