% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IsMissingSpecies}
\alias{IsMissingSpecies}
\title{Missing Species Vector}
\usage{
IsMissingSpecies(locus, SpeciesNames)
}
\arguments{
\item{locus}{A single locus (can have multiple sites)}

\item{SpeciesNames}{Vector of species names that will cluster individuals. This will likely be rownames(SNPdataset)}
}
\value{
Returns a TRUE/FALSE vector for each locus in the dataset
}
\description{
This function will determine if entire species are missing data (across all sites of a locus). Species names are important for this function, so if they are read in incorrectly, if could affect the results. They should be in the format where each species shares a unique flag and are then numbered (for example, species1, species2, species3 would be three individuals of the same species). If you want to check species see the function \code{GetSpecies}.
}
\examples{
data(fakeData)
Spnames <- rownames(fakeData)
IsMissingSpecies(fakeData[,1], Spnames)
IsMissingSpecies(fakeData[,2], Spnames)
IsMissingSpecies(fakeData[,3], Spnames)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{GetSpecies} \link{RemoveMissingSpeciesLoci}
}

