% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MakeMinIndTable}
\alias{MakeMinIndTable}
\title{Reduce Data by Minimum Individuals}
\usage{
MakeMinIndTable(SNPdataset, showEvery = 5, missingChar = "any",
  file = NULL, ...)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{showEvery}{An integer to describe how many datasets to calculate.  If \code{showEvery=3}, you will get a table that returns minimum individual data increasing by 3 individuals each time (ex: 1, 4, 7, 10, etc...)}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}

\item{file}{An optional agument to write the table to a file.}

\item{...}{Optional additional arguments to pass to file writing (such as appending the file with more data).}
}
\value{
Returns a matrix with the minimum number of individuals, the number of loci, variable sites, and percent of missing data for that number of individuals.
}
\description{
This function will reduce a dataset to include sites or loci that meet the minimum number of individuals.  For example, if you want only data for which at least 75% of the individuals in your dataset are represented, then it will remove those loci/sites that have less.
}
\examples{
data(fakeData)
MakeMinIndTable(fakeData, showEvery=1)
MakeMinIndTable(fakeData, showEvery=3)
}
\seealso{
\link{ReadSNP} \link{ReduceMinInd} \link{plotMissing}
}

