% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePresentAbsent.R
\name{MakePresentAbsent}
\alias{MakePresentAbsent}
\title{Make Data Present/Absent}
\usage{
MakePresentAbsent(SNPdataset, calc = "loci", missingChar = "any")
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci". If you choose to calculate missing data by loci, any site with information will count (for example, "NNNNA" will count as non-missing)}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}
}
\value{
Returns a matrix with each column representing a single site with present/absent data.
}
\description{
This function will remove breaks between loci and change all data to boolean values.  If data is present in any base pair or ambiguous character, then the value is assigned present (1), if data is missing (ie, "N", "-", or "?") then it is assigned absent (0).
}
\examples{
data(fakeData)
MakePresentAbsent(fakeData, "sites")
MakePresentAbsent(fakeData, "loci")
}
\seealso{
\link{ReadSNP} \link{ConvertMissingData} \link{CalculateMissingData}
}
