% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeStructureFormat.R
\name{MakeStructureFormat}
\alias{MakeStructureFormat}
\title{Make Structure Formatted Files}
\usage{
MakeStructureFormat(SNPdataset)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}
}
\value{
Returns a table with alleles.
}
\description{
This function will create tables that are needed as Structure input. Alleles are coded
numerically (A=1, C=2, G=3, T=4, NA=-9). Each sample's unlinked file gets reduced to 
its individual alleles (i.e. ambiguity codes are assumed to be heterozygous snps).
}
\examples{
data(fakeData)
fakeData <- RemoveNonBinary(RemoveInvariantSites(fakeData))
MakeStructureFormat(fakeData)
}
\seealso{
\link{ReadSNP} \link{AddAFlag} \link{ReturnAlleleCounts}
}
