% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercentMissing.R
\name{PercentMissing}
\alias{PercentMissing}
\title{Percent Missing Data}
\usage{
PercentMissing(locus, missingChar = "any")
}
\arguments{
\item{locus}{A single locus}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}
}
\value{
numerical, A single value for percentage of missing data (0 is none, 1 is all)
}
\description{
This function will determine the percent missing from a locus.
}
\examples{
data(fakeData)
PercentMissing(fakeData[1,1])
PercentMissing(fakeData[1,3])
}
\seealso{
\link{ReadSNP} \link{MakePresentAbsent} \link{CalculateMissingData} \link{IsMissing}
}
