\name{phrynomics-package}
\alias{phrynomics-package}
\alias{phrynomics}
\alias{Phrynomics}
\docType{package}
\title{phrynomics}
\description{
	A sweet set of SNP dataset manipulation tools
}
\details{
	\tabular{ll}{
	Package: \tab phrynomics\cr
	Type: \tab Package\cr
	Version: \tab 1.5 \cr
	Date: \tab 2014-6-16 \cr
	License: \tab GPL (>= 2)\cr}
}
\author{
	\tabular{ll}{
     Authors: \tab Barb Banbury & Adam Leache \cr
     Maintainer: \tab Barb Banbury <darwinthesun@gmail.com> \cr}
}
