% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrynomics.R
\docType{package}
\name{phrynomics}
\alias{Phrynomics}
\alias{phrynomics}
\alias{phrynomics-package}
\title{phrynomics - Phylogenetic Genomics on Phrynosomatid Lizards}
\description{
A sweet set of SNP dataset maipulation tools. Initially designed to modify our phrynosomatid lizard datasets, the new phrynomics version should be able to modify any SNP datasets. The functions will read in sequences, check for variable and binary sites, remove of taxa or sites, and transform characters in various ways. There are options for exporting data to fit several popular phylogenetics programs (ex: RAxML, MrBayes, and SNAPP).
}
\details{
\tabular{ll}{
Package: \tab phrynomics \cr
Type: \tab Package \cr
Version: \tab 1.5 \cr
Date: \tab 2014-07-21 \cr
License: \tab GPL(>= 2) \cr
}
}
\author{
Barb Banbury & Adam Leache
}

