% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.snp}
\alias{plot.snp}
\title{plot.snp}
\usage{
\method{plot}{snp}(x, calc = "loci", ...)
}
\arguments{
\item{x}{an object in the class "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci"}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
Plot SNP Matrix data
}
\details{
This function will create a three-panel plot, the first that shows the number of sites per locus, the second that shows the frequency distribution of sites per locus, and the third shows the amount of missing data . This doesn't work when there is a single locus, however you can plot the third panel alone by calling \code{plotMissing}.
}
\examples{
data(fakeData)
myData <- ReadSNP(fakeData)
plot(myData)
}
\seealso{
\link{ReadSNP} \link{summary.snp} \link{plotMissing}
}

