% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot SNP Missing Data Heatmap}
\usage{
plotHeatmap(SNPdataset, calc = "loci", col = "black",
  colorMissing = FALSE, grid = TRUE, xaxisTicks = 10, ...)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci". Calculating by "loci" will give a percentage of missing data for each cell.}

\item{col}{Optional color parameter that represents data (missing data will be white)}

\item{colorMissing}{If set to TRUE, then missing data will be colored and white will be complete.}

\item{grid}{Include background grid}

\item{xaxisTicks}{Number of tick marks for the x-axis}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
This function will create a heatmap plot of the level of missing data. Colors represent degrees of missing data, completely missing is opaque color.
}
\examples{
data(fakeData)
plotHeatmap(fakeData, calc="loci")
plotHeatmap(fakeData, calc="sites", "goldenrod")

exdata <- ReadSNP(system.file("extdata", "SNPdata.snps", package="phrynomics"))
plotHeatmap(exdata, calc="loci")
plotHeatmap(exdata, colorMissing=TRUE, grid=FALSE, col="red", xaxisTicks=0, cex.axis=0.5)
}
\seealso{
\link{ReadSNP} \link{CalculateMissingData} \link{PercentMissing}
}

