% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{plot.snp}
\usage{
plotMissing(SNPdataset, calc = "loci", ...)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci"}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
Plot SNP Matrix data
}
\details{
This function will create a plot of the amount of missing data. It is calculated as the number of retained loci or sites given the minimum taxa that must be included.  For example, if you want a dataset that only includes sites with all your individuals (ie NO missing data), you can see how many sites will be included.
}
\examples{
data(fakeData)
myData <- ReadSNP(fakeData)
plotMissing(myData, calc="loci")
plotMissing(myData, calc="sites")
}
\seealso{
\link{ReadSNP} \link{summary.snp} \link{plot.snp}
}

