% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.snp}
\alias{print.snp}
\title{print.snp}
\usage{
\method{print}{snp}(x, ...)
}
\arguments{
\item{x}{an object in the class "snp"}
}
\description{
Compact Display of SNP Matrix
}
\details{
This function prints a short summary of the SNP matrix to the console.
}
\examples{
a <- matrix(data=c("AANNTGG", "AATTTGC", "TAAATGC"), dimnames=list(c("A", "B", "C"), "locus"))
write.table(a, file="ex.data", col.names=FALSE)
b <- ReadSNP("ex.data")
print(b)
str(b)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{summary.snp}
}

