% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReadInterleavedNexus}
\alias{ReadInterleavedNexus}
\title{Read Interleaved Nexus File}
\usage{
ReadInterleavedNexus(file)
}
\arguments{
\item{file}{A file name specified by either a variable or a double-quoted string (ie, file=""). Also accepts R objects that are data frames or matrices and converts to the class "snp".}
}
\value{
Returns a data frame with rownames as species and column(s) as loci
}
\description{
Read Interleaved Nexus File
}
\details{
This function reads in SNP datasets that are interleaved and converts them to data frame. These can then be written as standard nexus or phylip formatted files.
}
\seealso{
\link{WriteSNP}
}

