% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadSNP.R
\name{ReadSNP}
\alias{ReadSNP}
\alias{snp}
\title{Read SNP}
\usage{
ReadSNP(file, row.names = 1, preprocess = TRUE, fileFormat = NULL,
  extralinestoskip = 0)
}
\arguments{
\item{file}{A file name specified by either a variable or a double-quoted string (ie, file=""). Also accepts R objects that are data frames or matrices and converts to the class "snp".}

\item{row.names}{A numeric indicating which column of data the row names occur (default is 1)}

\item{preprocess}{If preprocess=TRUE, it will remove underscores from pyrad output ("_")}

\item{fileFormat}{If NULL, phrynomics will try to guess whether your data is phylip or nexus formatted.  Otherwise you can specify, "phy" or "nex".}

\item{extralinestoskip}{Any blank lines or comments in the beginning of a file to skip}
}
\value{
Returns an object in the class "snp" with the following components:
\itemize{
\item \code{data} The data matrix
\item \code{ntax} Number of total taxa present in data matrix
\item \code{nloci} Number of loci present in the data matrix
\item \code{nsites} A vector of the number of sites present in each locus
}
}
\description{
Read a SNP File
}
\details{
This function reads in SNP datasets into R from nexus or phylip formatted files or converts an R object dataset into one of the class "snp".
}
\examples{
#Read in data from file
a <- matrix(data=c("AANNTGG", "AATTTGC", "TAAATGC"), dimnames=list(c("A", "B", "C"), "locus"))
write.table(a, file="ex.data", col.names=FALSE)
ReadSNP("ex.data")

#Read in data from an R object
data(fakeData)
ReadSNP(fakeData)
}
\seealso{
\link{WriteSNP}
}

