% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReduceMinInd}
\alias{ReduceMinInd}
\title{Reduce Data by Minimum Individuals}
\usage{
ReduceMinInd(SNPdataset, calc = "loci", threshold = 0.9,
  missingChar = "any", chatty = FALSE)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci". If you choose to calculate missing data by loci, any site with information will count (for example, "NNNNA" will count as non-missing)}

\item{threshold}{The minimum number of individuals that must be present for data to be retained.  This can be either a percent of individuals or a whole integer. When threshold=1, it means at least one individual, not 100\% of the data.}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}

\item{chatty}{Optional print to screen messages}
}
\value{
Returns a subset dataset.
}
\description{
This function will reduce a dataset to include sites or loci that meet the minimum number of individuals.  For example, if you want only data for which at least 75% of the individuals in your dataset are represented, then it will remove those loci/sites that have less.
}
\examples{
data(fakeData)
MakePresentAbsent(fakeData, "sites")
MakePresentAbsent(fakeData, "loci")
}
\seealso{
\link{ReadSNP} \link{MakeMinIndTable} \link{CalculateMissingData} \link{plotMissing}
}

