% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveGroups}
\alias{RemoveGroups}
\title{Remove Groups}
\usage{
RemoveGroups(SNPdataset, groupFlag = "")
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{groupFlag}{Unique flag for ingroup taxa}
}
\value{
Returns a subset dataset with only ingroup taxa
}
\description{
This function will remove taxon names that do not match with an ingroup ID.  For example, if all the ingroup taxa share a taxon flag identifier "PH", then this function will remove any that do not have this flag. Useful for comparing ingroup to whole data or for analyses like SNAPP that do not require ingroups.
}
\examples{
data(fakeData)
RemoveGroups(fakeData, groupFlag ="in")
}
\seealso{
\link{ReadSNP} \link{WriteSNP}
}

