% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ReturnAlleleCounts}
\alias{ReturnAlleleCounts}
\title{Return Allele COunts}
\usage{
ReturnAlleleCounts(site, allelenames = TRUE, bgc = FALSE)
}
\arguments{
\item{site}{Single SNP site as a vector}

\item{allelenames}{whether to return a named vector}

\item{bgc}{For Bayesian Genomic Clines (Gompert and Buerkle 2011)}
}
\value{
Returns a named list of allele counts.
}
\description{
This function will calculate the number of alleles in a population. Heterozygotes get split between their bases.
}
\examples{
data(fakeData)
splitfakeData <- SplitSNP(fakeData)
ReturnAlleleCounts(splitfakeData[,1])
ReturnAlleleCounts(splitfakeData[,2])
lapply(splitfakeData, ReturnAlleleCounts)
}
\seealso{
\link{ReadSNP} \link{GetBaseFrequencies} \link{ExportPops}
}

