% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnMrBayesAmbyCode.R
\name{ReturnMrBayesAmbyCode}
\alias{ReturnMrBayesAmbyCode}
\title{Return MrBayes Ambiguity Codes}
\usage{
ReturnMrBayesAmbyCode(NucCode)
}
\arguments{
\item{NucCode}{An ambiguity code}
}
\value{
Returns a character vector with numerical base possibilities surrounded by parentheses. Sites will be transformed where A=1, T=2, G=3, and C=4.
}
\description{
This function will take an IUPAC ambiguity code and return a set of bases in MrBayes parenthetical format
}
\examples{
ReturnMrBayesAmbyCode("A")
ReturnMrBayesAmbyCode("R")
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ReturnAmbyCode}
}
