% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SplitSNP}
\alias{SplitSNP}
\title{Split SNP}
\usage{
SplitSNP(SNPdataset)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}
}
\value{
Returns a matrix with a single SNP in each colomn. Empty columns represent breaks in loci.
}
\description{
SplitSNP breaks character strings into single character elements. For example, the string "ATA" will return "A" "T" "A". Breaks in loci are maintained.
}
\examples{
data(fakeData)
fakeData <- ReadSNP(fakeData)
SplitSNP(fakeData)
}
\seealso{
\link{ReadSNP} \link{cSNP}
}

