% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranslateBases.R
\name{TranslateBases}
\alias{TranslateBases}
\title{Translate Bases}
\usage{
TranslateBases(
  SNPdataset,
  translateMissing = TRUE,
  translateMissingChar = "-",
  ordered = FALSE
)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{translateMissing}{Logical. Will transform missing characters if TRUE}

\item{translateMissingChar}{The character missing data will transform to. This has limited functionality, in that any missing data character (ie, "N", "-", and "?") will be transformed to the named character. For more options see \code{ConvertMissingData}.}

\item{ordered}{Logical. If ordered = FALSE (default), Sites will be transformed where A=1, T=2, G=3, and C=4. In the case of ambiguous characters, the possible bases are returned in parentheses. This is the format for MrBayes MkV models. If ordered = TRUE, each site is evaluated for base frequencies and the base with the most frequent occurence is assigned a 0 and the less frequent base is assigned a 1.  THis only works for binary traits.  Any heterozygotes (ambiguity with the present two bases) are assigned a 1.  This is the format for SNAPP analyses.}
}
\value{
Returns a translated dataset.
}
\description{
This function will translate the bases from alpha characters to numeric. Several options exist for transformation, such as transforming missing characters and ordering characters by frequency.
}
\examples{
data(fakeData)

# Translate to MrBayes MkV formatting
TranslateBases(fakeData, translateMissing=FALSE, ordered=FALSE)
TranslateBases(fakeData, translateMissing=TRUE, translateMissingChar="?")

# Translate to SNAPP formatting
fakeData <- RemoveNonBinary(fakeData)
TranslateBases(fakeData, translateMissing=TRUE, ordered=TRUE)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{RemoveNonBinary}
}
