\name{WriteToPhy}
\alias{WriteToPhy}
\alias{WriteToPhy}
\title{Write Data To Phylip File}
\description{This function write SNP data to a phylip file}
\usage{
WriteToPhy(data, fileName="file.phy")
}
\arguments{
  \item{data}{SNP dataset}
  \item{fileName}{user defined file name}
}
\details{This will write the phylip file to the working directory.}
\value{Saves a phylip formatted file to the working directory.}
\author{B.Banbury}

