% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addazero.R
\name{addazero}
\alias{addazero}
\title{Add A Zero to alleleCounts list}
\usage{
addazero(alleleCounts, rownamesOriginal)
}
\arguments{
\item{alleleCounts}{allele count list out of ReturnAlleleCounts()}

\item{rownamesOriginal}{rownames from full dataset in ReturnAlleleCounts()}
}
\value{
Returns a new alleleCounts list with zeros in the right position
}
\description{
This function will add a zero if allele counts are missing from subpops
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ExportPops}
}

