% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CalculateMissingData}
\alias{CalculateMissingData}
\title{Calculate Amount of Missing Data}
\usage{
CalculateMissingData(SNPdataset, calc = "sites", missingChar = "any")
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}

\item{calc}{Calculate missing data either by individual "sites" or "loci". If you choose to calculate missing data by loci, any site with information will count (for example, "NNNNA" will count as non-missing)}

\item{missingChar}{The character that represents missing data, This function will search for missing data ("N", "-" or "?"). You can choose to leave the default "any" which will search for all three variables or select whichever are present in your dataset.}
}
\value{
Returns a vector of missing data for each site or locus
}
\description{
This function will calculate what percentage of the dataset is missing.  You must give it the missing character value (ex: "N", "-"), and you can not have more than one value in the dataset.
}
\examples{
data(fakeData)
CalculateMissingData(fakeData, calc="sites")
CalculateMissingData(fakeData, calc="loci")
}
\seealso{
\link{ReadSNP} \link{SplitSNP}
}

