% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ConvertMissingData}
\alias{ConvertMissingData}
\title{Convert Missing Data Character}
\usage{
ConvertMissingData(SNPdataset, oldmissing = c("-", "N"), newmissing = "N")
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{oldmissing}{The character that represents missing data}

\item{newmissing}{The new character to represent missing data}
}
\value{
Returns a dataset of the same dimentions as the original, but with swapped out characters.
}
\description{
This function will convert any character in a dataset to another.  This is helpful when one program outputs "N" representing missing data and the next requires "-".  You could also use this for converting other types of characters too (for example, if you wanted "A" to be "O").
}
\examples{
data(fakeData)
ConvertMissingData(fakeData, "N", "O")
ConvertMissingData(fakeData, c("N", "-"), "?")
ConvertMissingData(fakeData, "any", "?")
}
\seealso{
\link{ReadSNP} \link{CalculateMissingData}
}

