% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cSNP.R
\name{cSNP}
\alias{cSNP}
\title{Combine SNP}
\usage{
cSNP(splitSNP, KeepVector = NULL, maintainLoci = TRUE)
}
\arguments{
\item{splitSNP}{SNP data matrix in the class "data.frame", "matrix", or "snp", where each site is a separate column}

\item{KeepVector}{A boolean vector for returning specific sites.  If left to NULL, all SNPs will be returned.}

\item{maintainLoci}{If \code{maintainLoci=TRUE}, loci will be split by a space in the returned matrix.}
}
\value{
Returns a data frame of sites either with or without spaces indicating breaks in loci.
}
\description{
Combine split character strings into single character elements. For example, the string "A" "T" "A" will return "ATA". Breaks in loci are maintained using \code{maintainLoci=TRUE}.
}
\examples{
data(fakeData)
fakeData <- ReadSNP(fakeData)
cSNP(SplitSNP(fakeData))
cSNP(SplitSNP(fakeData), maintainLoci=FALSE)

coolData <- matrix(data=c("A", "T", "C", "A", "S", "N"), nrow=2)
cSNP(coolData)
}
\seealso{
\link{ReadSNP} \link{SplitSNP}
}

