\name{DataOverlap}
\alias{DataOverlap}
\title{Calculate amount of data overlap}
\description{This function will calculate the amount of datalap any given taxon has with the rest of the taxa in the dataset}
\usage{
DataOverlap(dataset)
}
\arguments{
  \item{dataset}{SNP dataset}
}
\details{}
\value{This will return a list of two items. The first is the present-absent data matrix with individual likelihood of overlap. The second is a vector of each taxon with average overlap among other taxa. 
}
\author{B.Banbury}
