% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FileFormat}
\alias{FileFormat}
\title{File Format}
\usage{
FileFormat(file)
}
\arguments{
\item{file}{A file name specified by either a variable or a double-quoted string (ie, file="")}
}
\value{
Returns either "phy" or "nex"
}
\description{
This is a function to autodetect phylip or nexus style file formatting.
}
\examples{
a <- matrix(data=c("AANNTGG", "AATTTGC", "TAAATGC"), dimnames=list(c("A", "B", "C"), "locus"))
WriteSNP(a, "example.phy")
FileFormat("example.phy")

WriteSNP(a, "example.nex", format="nexus")
FileFormat("example.nex")
}
\seealso{
\link{ReadSNP}
}

