\name{GetBaseFrequencies}
\alias{GetBaseFrequencies}
\title{Get Base Frequencies}
\description{This function will get base frequencies for a single SNP}
\usage{
GetBaseFrequencies(SNP)
}
\arguments{
  \item{SNP}{A single SNP site}
}
\details{}
\value{Returns a named vector with bases and frequencies
}
\author{B.Banbury}
