% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetBaseFrequencies}
\alias{GetBaseFrequencies}
\title{Get Base Frequencies}
\usage{
GetBaseFrequencies(site)
}
\arguments{
\item{site}{Single SNP site as a vector}
}
\value{
Returns a named vector of frequencies.
}
\description{
This function will calculate base frequencies for any single site. It will not return missing or ambiguous data, so values may not add up to one.
}
\examples{
data(fakeData)
splitfakeData <- SplitSNP(fakeData)
GetBaseFrequencies(splitfakeData[,1])
GetBaseFrequencies(splitfakeData[,2])
GetBaseFrequencies(splitfakeData[,3])
}
\seealso{
\link{ReadSNP} \link{ReturnUniqueBases}
}

