% Generated by roxygen2 (4.0.1): do not edit by hand
\name{IsBinary}
\alias{IsBinary}
\title{Is Binary}
\usage{
IsBinary(site)
}
\arguments{
\item{site}{A single SNP site, as a vector of bases}
}
\value{
Boolean response whether binary
}
\description{
This function will determine if a site is binary (TRUE) or not (FALSE). For example, c("A", "A", "T") will return TRUE; c("A", "A", "A") or c("A", "G", "T") will return FALSE. Ambiguity codes are taken into account to mean either heterozygotes or uncertainty. Such that c("A", "A", "R") will return TRUE; c("A", "A", "Y") will return FALSE.
}
\examples{
IsBinary(c("A", "A", "R"))
IsBinary(c("A", "A", "Y"))
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{IsVariable}
}

