% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IsVariable}
\alias{IsVariable}
\title{Is Variable}
\usage{
IsVariable(SNP)
}
\arguments{
\item{SNP}{A single SNP, as a vector of bases}
}
\value{
Boolean response of variability
}
\description{
This function will determine if a site is variable (TRUE) or not (FALSE). Ambiguity codes are taken into account to mean either heterozygotes or uncertainty. For example, c("A", "A", "S") will return TRUE, because "S" reduces to "G" and "C" and so either of those is variable with "A"; c("A", "A", "M") will return FALSE, because M can be either A or C. If the "M" is uncertain and is an "A" then it is not variable.
}
\examples{
IsVariable(c("A", "A", "A", "S"))
IsVariable(c("A", "A", "A", "M"))
}
\seealso{
\link{ReadSNP} \link{WriteSNP}
}

