\name{phrynomics-package}
\alias{phrynomics-package}
\alias{phrynomics}
\alias{Phrynomics}
\docType{package}
\title{phrynomics}
\description{
	A sweet set of SNP dataset manipulation tools
}
\details{
	\tabular{ll}{
	Package: \tab phrynomics\cr
	Type: \tab Package\cr
	Version: \tab 1.0 \cr
	Date: \tab 2014-4-29 \cr
	License: \tab GPL (>= 2)\cr}
}
\author{
Barb Banbury, Adam Leaché
Maintainer: Barb Banbury <darwinthesun@gmail.com>
}
