% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveInvariantSites.R
\name{RemoveInvariantSites}
\alias{RemoveInvariantSites}
\title{Remove Invariant Sites}
\usage{
RemoveInvariantSites(SNPdataset, chatty = FALSE)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{chatty}{Optional print to screen messages}
}
\value{
Returns a subset dataset with only variable sites (ie SNPs).
}
\description{
This function will determine if a site is variable (TRUE) or not (FALSE) and remove those sites that are not variable. Ambiguity codes are taken into account to mean either heterozygotes or uncertainty. For example, c("A", "A", "S") will return TRUE, because "S" reduces to "G" and "C" and so either of those is variable with "A"; c("A", "A", "M") will return FALSE, because M can be either A or C. If the "M" is uncertain and is an "A" then it is not variable.
}
\examples{
data(fakeData)
RemoveInvariantSites(fakeData)
RemoveInvariantSites(fakeData, chatty=TRUE)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{IsVariable}
}

