% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RemoveMissingSpeciesLoci.R
\name{RemoveMissingSpeciesLoci}
\alias{RemoveMissingSpeciesLoci}
\title{Remove Loci That Do Not have Species Coverage}
\usage{
RemoveMissingSpeciesLoci(SNPdataset, chatty = FALSE)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{chatty}{Optional print to screen messages}
}
\value{
Returns a subset dataset with sites where at least one individual has data.
}
\description{
This function will remove loci where entire species are missing data (across all sites of a locus). Species names are important for this function, so if they are read in incorrectly, if could affect the results. They should be in the format where each species shares a unique flag and are then numbered (for example, species1, species2, species3 would be three individuals of the same species). If you want to check species see the function \code{GetSpecies}.
}
\examples{
data(fakeData)
RemoveMissingSpeciesLoci(fakeData)
RemoveMissingSpeciesLoci(fakeData, chatty=TRUE)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{GetSpecies} \link{IsMissingSpecies}
}

