% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RemoveNonBinary}
\alias{RemoveNonBinary}
\title{Remove Non-binary Sites}
\usage{
RemoveNonBinary(SNPdataset, chatty = FALSE)
}
\arguments{
\item{SNPdataset}{SNP data in the class "matrix", "data.frame", or "snp"}

\item{chatty}{Optional print to screen messages}
}
\value{
Returns a subset dataset with only variable sites.
}
\description{
This function will determine if a site is binary (TRUE) or not (FALSE) and remove those sites that are not binary.  For example, c("A", "A", "T") will be kept; c("A", "A", "A") or c("A", "G", "T") will be removed. Ambiguity codes are taken into account to mean either heterozygotes or uncertainty.
}
\examples{
data(fakeData)
RemoveNonBinary(fakeData)
RemoveNonBinary(fakeData, chatty=TRUE)
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{RemoveInvariantSites} \link{IsBinary}
}

