% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnMrBayesCode.R
\name{ReturnMrBayesCode}
\alias{ReturnMrBayesCode}
\title{Return MrBayes Codes}
\usage{
ReturnMrBayesCode(base)
}
\arguments{
\item{base}{A nucleotide base}
}
\value{
Returns a single number that corresponds with a base pair
}
\description{
This function will take a nucleotide base and return a corresponding number
}
\examples{
ReturnMrBayesCode("A")
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ReturnAmbyCode} \link{ReturnMrBayesAmbyCode}
}

