% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReturnNucs}
\alias{ReturnNucs}
\title{Return Ambiguity Codes}
\usage{
ReturnNucs(NucCode, forSNAPP = FALSE)
}
\arguments{
\item{NucCode}{An ambiguity code}

\item{forSNAPP}{Logical. If FALSE (default), then missing data characters will be returned as all possibilities. If TRUE, the return for missing data will be returned "-".}
}
\value{
Returns a character vector with base possibilities.
}
\description{
This function will take an IUPAC ambiguity code and return a set of bases
}
\examples{
ReturnNucs("N", forSNAPP=FALSE)
ReturnNucs("N", forSNAPP=TRUE)
ReturnNucs("K")
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ReturnAmbyCode}
}

