% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReturnUniqueBases}
\alias{ReturnUniqueBases}
\title{Return Unique Bases}
\usage{
ReturnUniqueBases(SNP)
}
\arguments{
\item{SNP}{A single SNP}
}
\value{
Returns a character vector with base possibilities.
}
\description{
This function will return the unique nucleotide bases of an SNP accounting for an ambiguity code
}
\examples{
ReturnUniqueBases(c("A", "A", "R"))

data(fakeData)
fakeData <- SplitSNP(fakeData)
ReturnUniqueBases(fakeData[,1])
ReturnUniqueBases(fakeData[,12])
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{ReturnNucs}
}

