% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TakeSingleSNPfromEachLocus}
\alias{TakeSingleSNPfromEachLocus}
\title{Take Single SNP From Each Locus}
\usage{
TakeSingleSNPfromEachLocus(SNPdataset)
}
\arguments{
\item{SNPdataset}{SNP dataset in the class "matrix", "data.frame", or "snp"}
}
\value{
Returns a list of three items.  The first is the subset dataset with only a single site per locus. The second is a vector of the position of the chosen site within the locus. The third is a vector that details if the site was chosen based on maximum coverage ("M") or randomly ("R").
}
\description{
This function will keep a single site from each locus in the dataset.  For example, if your dataset contains 450 sites in 3 loci, your final dataset will only include 3 sites. This is to prevent linkage issues across sites within the same locus. This function will initially caluclate the percentage of species that have data for each site within a locus and it will choose to keep the site with the most data.  If there is equally good coverage across sites, then one is chosen by random.
}
\examples{
data(fakeData)
TakeSingleSNPfromEachLocus(fakeData)
TakeSingleSNPfromEachLocus(fakeData)[[1]]
}
\seealso{
\link{ReadSNP} \link{WriteSNP} \link{CalculateMissingData}
}

